ComponentRegistry will not throw an error when redefining same component
If the registry is restored from the persistance database, it will contain
all component definitions already, but the plugins will still try to
re-create them. We should not throw an exception in this case, however, we
should throw an exception should a different layout be provided or if the
owner GUID has changed (the same layout but with different semantics may be
registered by other plugin).