Comitting changes to projects and solution made by MonoDevelop
These are introduced by the MonoDevelop even when no changes to the project
are done. Perhaps MonoDevelop is more strict about file format, so it
reformats it automatically. To avoid this situation in the future, new
projects should be created from MonoDevelop (Xamarin Studio) or opened by
it at least once before adding to repository.