Removed the concept of upgrading attributes.
This has not been used so far and upgrading can still be simulated by
registering a new component with a different name (e.g. location2) and
updateing all entities in the world. If this becomes common, we can still
write a new method in the Core that will automate this.