Close the service host as soon as the server is ready.
This is require to be able to start two servers. Second server also needs
a service host to report when it's ready, but can't use the same service
URI. Therefore we close the service host asap. Later if Testing interface
is used for something else than reporting when the server is ready, a
better solution may be required (e.g. based on a config file with a service
URI).